refactor(toast): changed components structure
