+++ src/components/Notification/Notification.stories.mdx
@@ -2,7 +2,7 @@ import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
 import { withKnobs, text, select, boolean } from '@storybook/addon-knobs';
 import Notification from '../Notification';
 import Toast from '../Notification/Toast';
-import NotificationVisual from '../Notification/subcomponents/NotificationVisual';
+import NotificationVisual from '../Notification/NotificationVisual';
 import NotificationsContainer from '../Notification/subcomponents/NotificationsContainer';
 import PresentComponent from '../storyUtils/PresentComponent';
 import Stack from '../storyUtils/Stack';
